Factor out the window manager to separate file, rename to EditorWindows
